// src/app/common/models/100news.models.ts
export class News {
    id?: Number;
    typeid?: string;
    siteid?: string;
    title?: string;
    url?: string;
    postdate?: any;
    changedate?: string;
    summary?: string;
    content?: string;
    pathimage?: string;
    author?: string;
    approved?: any;
    comment?: string;
    lang?: string;
    isfocus?: any;
    isNew?: string;
    view?: string;
}